schema overhaul: create customers and customer_files, implementing .hasTable & .createTable for all schema

+++ src/db/migrations/20240307081022_create_users_table.cjs
@@ -2,20 +2,24 @@
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.up = function(knex) {
-  return knex.schema.createTable('users', (table) => {
-    table.increments('id').primary().unique().notNullable();
-    table.string('email').unique().notNullable();
-    table.string('password').notNullable();
-    table.timestamp('created_at').defaultTo(knex.fn.now());
-    table.timestamp('updated_at').defaultTo(knex.fn.now());
-  })
+exports.up = async(knex) => {
+  if (! await knex.schema.hasTable('users')) {
+    return knex.schema.createTable('users', (table) => {
+      table.increments('id').primary().unique().notNullable();
+      table.string('email').unique().notNullable();
+      table.string('password').notNullable();
+      table.timestamp('created_at').defaultTo(knex.fn.now());
+      table.timestamp('updated_at').defaultTo(knex.fn.now());
+    })
+  }
 };
 
 /**
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = function(knex) {
-  return knex.schema.dropTable('users')
+exports.down = async(knex) => {
+  if (await knex.schema.hasTable('users')) {
+    return knex.schema.dropTable('users')
+  }
 };
