+++ src/db/migrations/20240402200539_create_sessions_table.cjs
@@ -2,18 +2,22 @@
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.up = function(knex) {
-  return knex.schema.createTable('sessions', (table) => {
-    table.string('sid').primary().unique().notNullable();
-    table.json('sess').notNullable();
-    table.string('expire').notNullable();
-  })
+exports.up = async(knex) => {
+  if (! await knex.schema.hasTable('sessions')) {
+    return knex.schema.createTable('sessions', (table) => {
+      table.string('sid').primary().unique().notNullable();
+      table.json('sess').notNullable();
+      table.string('expire').notNullable();
+    })
+  }
 };
 
 /**
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = function(knex) {
-  return knex.schema.dropTable('sessions')
+exports.down = async(knex) => {
+  if (await knex.schema.hasTable('sessions')) {
+    return knex.schema.dropTable('sessions')
+  }
 };
